dynamixel - target reached threshold : bug precision

+++ src/linux/tools/robot_interface.cxx
@@ -779,7 +779,7 @@ int RobotInterface::dynamixel_set_max_torque(int dynamixel_type, uint8_t id, uin
 	return dynamixel_cmd(DYNAMIXEL_CMD_SET_MAX_TORQUE, dynamixel_type, id, val / 100.0f);
 }
 
-int RobotInterface::dynamixel_set_target_reached_threshold(int dynamixel_type, uint8_t id, uint8_t val)
+int RobotInterface::dynamixel_set_target_reached_threshold(int dynamixel_type, uint8_t id, float val)
 {
 	return dynamixel_cmd(DYNAMIXEL_CMD_SET_TARGET_REACHED_THRESHOLD, dynamixel_type, id, val);
 }
